Move things into a cli umbrella
