+++ src/cli/mod.rs
@@ -1,0 +1,8 @@
+/// CLI Indexes: Models
+pub mod models;
+
+/// CLI Indexes: Pipelines
+pub mod pipelines;
+
+/// CLI Indexes: Datasets
+pub mod datasets;
