+++ src/cli/models/bert/mod.rs
@@ -1,0 +1,7 @@
+/// CLI indexes: BERT model basics
+pub mod model;
+
+/// CLI indexes: BERT for Sequence Classification
+pub mod sequence_classification;
+
+pub use model::Model;
