+++ src/cli/models/mod.rs
@@ -1,0 +1,7 @@
+/// CLI Indexes: Base Model trait
+pub mod model;
+
+/// CLI Indexes: BERT models
+pub mod bert;
+
+pub use model::Model;
