+++ src/cli/pipelines/mod.rs
@@ -1,0 +1,7 @@
+/// CLI Indexes: Pipelines
+pub mod pipeline;
+
+/// CLI Indexes: Text Classification Pipeline
+pub mod text_classification;
+
+pub use pipeline::Pipeline;
