+++ src/lib.rs
@@ -13,6 +13,9 @@ pub mod datasets;
 /// Utilities
 pub mod utils;
 
+/// CLI indexes and utilities
+pub mod cli;
+
 /// Error macros
 #[macro_use]
 extern crate anyhow;
