+++ src/pipelines/text_classification/mod.rs
@@ -14,11 +14,3 @@ pub use batcher::{Batcher, Item};
 pub use inference::infer;
 pub use model::{Config, Model, ModelConfig};
 pub use training::train;
-
-use crate::models::{self, bert};
-
-/// The unique string token that identifies this pipeline
-pub static PIPELINE: &str = "text-classification";
-
-/// The default model to use for text classification
-pub static DEFAULT_MODEL: models::Model = models::Model::Bert(bert::BERT_BASE_UNCASED);
